requirements.txt
flask
mysql-connector-python
selenium
webdriver-manager
pandas
openpyxl
